Add integration tests for MembersDetailsApiMapper
